+++ src/components/ResultView.tsx
